requirements/base.txt
requests
aiohttp
selenium
Pillow
pyyaml
pytest
